Methods and Results - Completed, please check for word count or errors in code.
Completed the methods and results section. Please check the code for errors and go through it before writing the discussion.

Check the word count - from what I've checked, 1450 words currently, so discussion is approx 550 words (or longer if you deem some areas of the methods/results redundant).

Also, please refer to the figures in the discussion. If you need to, feel free to add figures to the data frame tables.